backend/requirements.txt
flask
pillow
pycryptodome
